+++ src/nodes/Hudu/optionLoaders/asset_layouts/index.ts
@@ -2,4 +2,3 @@ export * from './getAssetLayouts';
 export * from './getAssetLayoutFields';
 export * from './getCustomFieldsLayoutFields';
 export * from './getAssetLayoutFieldValues';
-export * from './getAssetLinkFields';